Support more ITAG clones
Some ITAG clones have different padding of ITAG name (trailing spaces,
tab characters, etc.). Also ITAG name can be in different case.
So, it is needed to support all of this variation.
Also some clones does not support linkLossAlert service, so it is
needed to skip writing data to it.